Package the GIST family into blackjax/mcmc/composed/ (pure move)
Finding: gist.py (the general seam), gist_step_size.py, and
gist_trajectory_length.py lived as three flat siblings in blackjax/mcmc/,
with no shared home for the composed step-size x trajectory-length instance
this module is about to grow (Issue#1066/#1059's h-then-L instance and its
verdict-routing entry). blackjax/adaptation/meta/ already establishes the
single-folder-package + back-compat-shim pattern for exactly this situation
(PR #1001).

Fix: moved the three files into blackjax/mcmc/composed/ (_seam.py, step_size.py,
trajectory_length.py), updating only their intra-module import of the general
seam (gist -> seam alias) and their Sphinx :mod: cross-references -- no other
line changed. Added composed/__init__.py with the package's public re-exports
and a module docstring stating the design (general seam + two shipped
instances; the composed h-then-L instance arrives in a follow-up change),
citing the GIST survey and adaptNUTS (arXiv:2408.08259) as the published
sources of the composition machinery per the cite-don't-claim convention.
The three old paths (blackjax/mcmc/gist*.py) become thin back-compat shims
that fire a DeprecationWarning (stacklevel=1) and re-export the full previous
surface, mirroring meta_adaptation.py's shim exactly. blackjax/__init__.py's
registrations now import from the new paths; public API names
(blackjax.gist_step_size, blackjax.gist_trajectory_length) are unchanged.
pytest.ini gets the matching filterwarnings entries for the three new shims.

Verified: import parity (old-path symbols are old_module.X is new_module.X);
both existing test files (42 tests, still importing the old paths through the
shim) pass unmodified; pre-commit clean.
