+++ pytest.ini
@@ -3,3 +3,6 @@ testpaths= "tests"
 filterwarnings =
     error
     ignore::DeprecationWarning:blackjax.adaptation.meta_adaptation
+    ignore::DeprecationWarning:blackjax.mcmc.gist
+    ignore::DeprecationWarning:blackjax.mcmc.gist_step_size
+    ignore::DeprecationWarning:blackjax.mcmc.gist_trajectory_length
